remove unused cache helpers
